+++ src/steps.ts
@@ -4,6 +4,7 @@ import { createHtmlReport } from 'axe-html-reporter';
 import { MemoryValue } from '@qavajs/core';
 import { readFile, writeFile, mkdir } from 'node:fs/promises';
 import { parse } from 'node:path';
+import any = jasmine.any;
 
 function htmlAttachment(results: AxeResults) {
     const reportHTML = createHtmlReport({
@@ -17,23 +18,32 @@ function htmlAttachment(results: AxeResults) {
 
 type World = { config: any, wdio?: any, playwright?: any, attach: (attachment: any, mime: string) => void };
 
-async function audit(world: World, options?: RunOptions & { saveAs?: string }) {
+async function audit(world: World, options?: RunOptions & { context?: any, saveAs?: string }) {
     if (!world.wdio && !world.playwright) throw new Error('Browser instance does not exist! Make sure that webdriverio or playwright steps are installed');
     const axeCode = await readFile('node_modules/axe-core/axe.min.js', { encoding: 'utf8' });
     let results;
+    const context = options?.context;
     if (world.playwright) {
         await world.playwright.page.evaluate(axeCode);
-        // @ts-ignore
-        results = await world.playwright.page.evaluate(options => window.axe.run(options), options);
+        results = await world.playwright.page.evaluate(({ context, options }: any) => {
+            const accContext = context ?? document;
+            // @ts-ignore
+            return window.axe.run(accContext, options);
+        }, { context, options });
     }
     if (world.wdio) {
         await world.wdio.browser.execute(axeCode);
-        // @ts-ignore
-        results = await world.wdio.browser.executeAsync((options, done) => window.axe.run(options).then(done), options);
+        results = await world.wdio.browser.executeAsync(({ context, options }: any, done: any) => {
+            const accContext = context ?? window.document;
+            // @ts-ignore
+            return window.axe.run(accContext, options).then(done)
+        }, { context, options });
     }
     if (options?.saveAs) {
         const path = parse(options.saveAs);
-        await mkdir(path.dir, { recursive: true });
+        if (path.dir) {
+            await mkdir(path.dir, { recursive: true });
+        }
         await writeFile(options.saveAs, JSON.stringify(results));
     }
     world.attach(htmlAttachment(results), 'base64:text/html');
@@ -67,7 +77,6 @@ When('I perform accessibility check:', async function (this: World, optionsMulti
     const options = JSON.parse(optionsMultiline);
     const results = await audit(this, options);
     if (results.violations.length > 0) {
-        console.log(results)
         throw new Error(`Accessibility check failed! Found ${results.violations.length} violations`);
     }
 });